astprocd: Use a list of warning messages
common/bundle: Return a list of messages rather than an optional string

astprocd: Handle a list of messages rather that splitting an optional
string by newline

This fixes an issue where spurious newlines would be printed in the
logfile! It's also generally a better approach as there is only a single
type that is returned and handled.