astctl: Correctly exit from loop in usercode log
This is another fix for #47, as the fix implemented in #86 didn't fix
the underlying cause.

The underlying issue is caused by the co-routine for `wait_broadcast`
sticking around after `halt` is called, meaning that the program never
exits. This issue likely exists in sr-robot3 also and is probably the
cause of https://github.com/srobo/sr-robot3/issues/32