astmetad: Remove bundle.toml from code bundle
We decided that we no longer want these in Slack and will replace them
with a replacement file that lives outside of the robot.zip. This will
make it significantly easier to create code bundles as no specialist
software will be required.

Whilst we lose the feature of warning the users that their robot may be
out of date, this was always a blue sky feature anyway and is probably
not of major loss to the user.