astmetad: Remove kit version and name from config
Now that the bundle.toml has been removed from the code bundle format,
these values serve little purpose other than being available to the user
in the Metadata dictionary.

This was always a little impractical to update as it means that we'd
need to update /etc/astoria.toml every time we did a new build of the
kit. It's likely to be much easier to print `/etc/issue` or similar