Add usercode_entrypoint to robot-settings.toml
This adds a usercode_entrypoint entry to the robot settings file for the purpose of determining which Python file is used as the entrypoint for the users' code.

This allows for users to change which file is executed within their usercode bundle.

The default can be manipulated in astoria.toml, so that a vendor can choose what filename will be usually executed.