Use pydantic.parse_obj_as instead of BaseModel(**data)
This commit changes how we instantiate pydantic BaseModel objects and thus how we validate the data that astoria is handling.

Previously, we were unpacking the data into kwargs and passing it into the constructor: `MyModel(**data)`, which works fine in the case that the `data` object is a dictionary. However, we can also load other data types from JSON and TOML, such as lists:

```python
>>> a = [1, 2, 3]
>>> dict(**a)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: dict() argument after ** must be a mapping, not list
```

Hence, there are potential runtime issues if we use this mechanism to create objects.

This commit changes to using `pydantic.parse_obj_as`, which safely handles objects that are not dictionaries.

Functionally, the code performs almost exactly the same.