Use tomli and tomli-w instead of toml library
The toml library is unmaintained and has known bugs.

Additionally, the tomli library is being integrated into the stdlib in Python 3.11 (PEP 680), making it the best choice of toml library. When we drop support for Python 3.10 and below, we will no longer need tomli as a dependency.

It should be noted that tomli is a load-only TOML library, and we need a separate library (tomli-w) to generate TOML files.

Fixes #125