astmetad: Generate error file on bad robot-settings.toml
When a robot-settings.toml file is bad, generate a robot-settings-error.txt
file on the USB drive also. This file contains a message to tell the user
that their settings has been regenerated, a detailed error for why the settings
were invalid and in most cases a copy of the old settings file.

An error file is not generated if the file is missing, as this is the expected
behaviour for a new USB drive.